feat: working oauth

+++ src/appState.ts
@@ -13,15 +13,6 @@ const DEFAULT_APP_STATE: AppState = {
   apiKeys: {},
 };
 
-const home = process.env.HOME || process.env.USERPROFILE;
-
-if (!home) {
-  throw new Error("Could not find home directory");
-}
-
-const STORAGE_PATH = path.join(home, ".algolia-mcp");
-const APP_STATE_PATH = path.join(STORAGE_PATH, "state.json");
-
 let SINGLETON: AppStateManager | null = null;
 
 abstract class StateManager {
@@ -36,6 +27,15 @@ export class AppStateManager implements StateManager {
   appState: AppState;
 
   static async load() {
+    const home = process.env.HOME || process.env.USERPROFILE;
+
+    if (!home) {
+      throw new Error("Could not find home directory");
+    }
+
+    const STORAGE_PATH = path.join(home, ".algolia-mcp");
+    const APP_STATE_PATH = path.join(STORAGE_PATH, "state.json");
+
     if (!SINGLETON) {
       await mkdir(STORAGE_PATH, { recursive: true });
 
@@ -54,7 +54,15 @@ export class AppStateManager implements StateManager {
   }
 
   static async reset(): Promise<void> {
+    const home = process.env.HOME || process.env.USERPROFILE;
+
+    if (!home) {
+      throw new Error("Could not find home directory");
+    }
+
     SINGLETON = null;
+
+    const STORAGE_PATH = path.join(home, ".algolia-mcp");
 
     await fs.rmdir(STORAGE_PATH, { recursive: true });
   }
@@ -68,6 +76,15 @@ export class AppStateManager implements StateManager {
   }
 
   async update(update: Partial<AppState>) {
+    const home = process.env.HOME || process.env.USERPROFILE;
+
+    if (!home) {
+      throw new Error("Could not find home directory");
+    }
+
+    const STORAGE_PATH = path.join(home, ".algolia-mcp");
+    const APP_STATE_PATH = path.join(STORAGE_PATH, "state.json");
+
     this.appState = { ...this.appState, ...update };
     await fs.writeFile(APP_STATE_PATH, JSON.stringify(this.appState));
   }
